+++ src/main/java/com/agenticcp/core/domain/cloud/adapter/outbound/aws/AwsAccountValidationAdapter.java
@@ -1,9 +1,9 @@
 package com.agenticcp.core.domain.cloud.adapter.outbound.aws;
 
 import com.agenticcp.core.domain.cloud.adapter.outbound.aws.config.AwsClientConfig;
-import com.agenticcp.core.domain.cloud.dto.AccountValidationRequest;
-import com.agenticcp.core.domain.cloud.dto.AccountValidationResult;
-import com.agenticcp.core.domain.cloud.dto.ConnectionTestResult;
+import com.agenticcp.core.domain.cloud.port.model.account.AccountValidationRequest;
+import com.agenticcp.core.domain.cloud.port.model.account.AccountValidationResult;
+import com.agenticcp.core.domain.cloud.port.model.account.ConnectionTestResult;
 import com.agenticcp.core.domain.cloud.port.outbound.AccountValidationPort;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
