+++ src/main/java/com/agenticcp/core/domain/cloud/mapper/CredentialCommandMapper.java
@@ -1,9 +1,6 @@
 package com.agenticcp.core.domain.cloud.mapper;
 
-import com.agenticcp.core.domain.cloud.command.DeleteCredentialCommand;
-import com.agenticcp.core.domain.cloud.command.ResolveCredentialCommand;
-import com.agenticcp.core.domain.cloud.command.StoreCredentialCommand;
-import com.agenticcp.core.domain.cloud.dto.RegisterCloudAccountRequest;
+import com.agenticcp.core.domain.cloud.port.model.account.*;
 import com.agenticcp.core.domain.cloud.entity.CloudProvider.ProviderType;
 import org.springframework.stereotype.Component;
