+++ src/main/java/com/agenticcp/core/domain/cloud/port/model/account/DeleteCredentialCommand.java
@@ -1,23 +1,18 @@
-package com.agenticcp.core.domain.cloud.command;
+package com.agenticcp.core.domain.cloud.port.model.account;
 
 import com.agenticcp.core.domain.cloud.entity.CloudProvider.ProviderType;
 import lombok.Builder;
 import lombok.Value;
 
 /**
- * 자격증명 조회 커맨드
+ * 자격증명 삭제 커맨드
  * 
  * @author AgenticCP Team
  * @version 1.0.0
  */
 @Value
 @Builder
-public class ResolveCredentialCommand {
-    
-    /**
-     * 테넌트 키
-     */
-    String tenantKey;
+public class DeleteCredentialCommand {
     
     /**
      * 프로바이더 타입
@@ -25,8 +20,8 @@ public class ResolveCredentialCommand {
     ProviderType providerType;
     
     /**
-     * 계정 범위 (AccountId, SubscriptionId, ProjectId)
+     * 자격증명 키 참조 (UUID)
      */
-    String accountScope;
+    String credentialKey;
 }
