+++ src/main/java/com/agenticcp/core/domain/cloud/port/outbound/AccountValidationPort.java
@@ -1,8 +1,8 @@
 package com.agenticcp.core.domain.cloud.port.outbound;
 
-import com.agenticcp.core.domain.cloud.dto.AccountValidationRequest;
-import com.agenticcp.core.domain.cloud.dto.ConnectionTestResult;
-import com.agenticcp.core.domain.cloud.dto.AccountValidationResult;
+import com.agenticcp.core.domain.cloud.port.model.account.AccountValidationRequest;
+import com.agenticcp.core.domain.cloud.port.model.account.ConnectionTestResult;
+import com.agenticcp.core.domain.cloud.port.model.account.AccountValidationResult;
 
 import java.util.Map;
