+++ src/test/java/com/agenticcp/core/domain/cloud/adapter/outbound/aws/AwsAccountValidationAdapterTest.java
@@ -2,9 +2,9 @@ package com.agenticcp.core.domain.cloud.adapter.outbound.aws;
 
 import com.agenticcp.core.common.exception.BusinessException;
 import com.agenticcp.core.domain.cloud.adapter.outbound.aws.config.AwsClientConfig;
-import com.agenticcp.core.domain.cloud.dto.AccountValidationRequest;
-import com.agenticcp.core.domain.cloud.dto.AccountValidationResult;
-import com.agenticcp.core.domain.cloud.dto.ConnectionTestResult;
+import com.agenticcp.core.domain.cloud.port.model.account.AccountValidationRequest;
+import com.agenticcp.core.domain.cloud.port.model.account.AccountValidationResult;
+import com.agenticcp.core.domain.cloud.port.model.account.ConnectionTestResult;
 import com.agenticcp.core.domain.cloud.entity.CloudProvider.ProviderType;
 import com.agenticcp.core.domain.cloud.exception.CloudErrorCode;
 import org.junit.jupiter.api.BeforeEach;
